+++ AkuariumItem.java
@@ -1,9 +1,9 @@
 import java.lang.Math;
 public abstract class AkuariumItem {
     /*! \brief AkuariumItem x-coordinate.*/
-    protected double x;
+    protected float x;
     /*! \brief AkuariumItem y-coordinate.*/
-    protected double y;
+    protected float y;
     /*! \brief Aquarium Pointer.*/
     protected Akuarium A;
 
@@ -14,29 +14,29 @@ public abstract class AkuariumItem {
     }
 
     /*! \brief AkuariumItem Constructor.*/
-    public AkuariumItem(double x, double y, Akuarium A) {
+    public AkuariumItem(float x, float y, Akuarium A) {
         this.x = x;
         this.y = y;
         this.A = A;
     }
 
     /*! \brief Set AkuariumItem x-coordinate.*/
-    public void setX(double val) {
+    public void setX(float val) {
         x = val;
     }
 
     /*! \brief Set AkuariumItem y-coordinate.*/
-    public void setY(double val) {
+    public void setY(float val) {
         y = val;
     }
 
     /*! \brief Return AkuariumItem x-coordinate.*/
-    public double getX() {
+    public float getX() {
         return x;
     }
 
     /*! \brief Return AkuariumItem y-coordinate.*/
-    public double getY() {
+    public float getY() {
         return y;
     }
 
@@ -44,8 +44,8 @@ public abstract class AkuariumItem {
     public abstract void ProsesWaktu();
 
     /*! \brief Return the Euclidean-distance of two AkuariumItem.*/
-    public static double jarak(AkuariumItem a1, AkuariumItem a2) {
-        return (double) Math.sqrt(Math.pow(a1.x - a2.x, 2) + Math.pow(a1.y - a2.y, 2));
+    public static float jarak(AkuariumItem a1, AkuariumItem a2) {
+        return (float) Math.sqrt(Math.pow(a1.x - a2.x, 2) + Math.pow(a1.y - a2.y, 2));
     }
 
 }
